pyproject.toml
[project]
name = "agent_sofia_backend"
version = "0.1.0"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "supabase",
    "python-dotenv",
    "livekit-agents>=1.3.3",
    "livekit-plugins-silero>=1.3.3",
    "livekit-plugins-turn-detector>=1.3.3",
    "openai>=2.8.1",
    "opencv-python-headless",
    "numpy",
]
